LPS-179094 Extract ScreenNavigationEntry methods from ObjectDefinitionsValidationsScreeNavigationCategory to ObjectDefinitionsValidationsScreeNavigationEntry, implement getCategoryKey in ObjectDefinitionsValidationsScreeNavigationEntry and make ObjectDefinitionsValidationsScreeNavigationCategory extends BaseObjectDefinitionsScreenNavigationCategory
